Solved 10055

+++ platforms/UHunt/answers/Problem10055.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int n, k;
+    long int n, k;
     while (cin >> n >> k)
         cout << (k > n ? (k - n) : (n - k)) << endl;
     return 0;
